Reorganize project - modularized - change pdfbox version to latest 1.8.x relase, so it match now the mustang sample
