+++ pdfa3-sample/src/main/resources/Ubuntu-R.ttf
